+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * Can integrate directly with Discord to post messages to a channel, using a user-specified Discord webhook.
 * Prefix positive INF with '+'.
 * Mission INF is now manually editable as well as automatically updated.
+* 'Select all' / 'Select none' checkbox at the top of each system to quickly enable / disable all factions for a system.
 
 
 ## v1.3.0 - 2021-09-06